ai-team-service/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
opencv-python-headless>=4.5.0
flask>=2.0.0
pillow>=8.0.0
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.0
scipy>=1.7.0
seaborn>=0.11.0
matplotlib>=3.4.0
tqdm>=4.62.0